requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
pydantic==2.4.2
pydantic-settings==2.0.3